LİBFT%100COMPLETED

+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: maygen <maygen@student.42istanbul.com.t    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/16 15:01:40 by maygen            #+#    #+#             */
-/*   Updated: 2022/10/21 11:07:38 by maygen           ###   ########.fr       */
+/*   Updated: 2022/10/25 21:43:21 by maygen           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (s)
